Moved analysis logic into its own `analysis` directory
