+++ src/selection/greedy.py
@@ -1,6 +1,6 @@
+from src.analysis.error_matrix import ErrorMatrix
 from src.animation.animation import Animation
 from src.animation.timeline import Timeline
-from src.selection.error_matrix import ErrorMatrix
 from src.selection.selection import Selection
 from src.selection.selector import Selector
