+++ src/selection/salient.py
@@ -1,8 +1,8 @@
 from typing import Dict, Tuple
 
+from src.analysis.error_matrix import ErrorMatrix
 from src.animation.animation import Animation
 from src.animation.timeline import Timeline
-from src.selection.error_matrix import ErrorMatrix
 from src.selection.selection import Selection
 from src.selection.selector import Selector
 from src.types import Time
